+++ frontend/components/take-away-page/component/schema/validationSchema.ts
@@ -1,6 +1,6 @@
 import * as zod from 'zod';
 import { Restaurant } from '@/types';
-import { validateBlacklist, validateOperatingTime } from '@/components/common/utils/validationUtils';
+import { isValidTime } from '@/components/common/utils/timeUtils';
 
 export const getFormDataSchema = (restaurant: Restaurant) => {
   const phoneSchema = zod
@@ -19,7 +19,27 @@ export const getFormDataSchema = (restaurant: Restaurant) => {
       notes: zod.string().optional(),
     })
     .superRefine((data, context) => {
-      validateBlacklist(data.phone, restaurant, context);
-      validateOperatingTime(data.date, data.time, restaurant, context);
+      const isTimeValid = isValidTime(
+        data.date,
+        data.time,
+        restaurant.Weekdaytime,
+        restaurant.Weekandtime,
+      );
+
+      if (!isTimeValid) {
+        context.addIssue({
+          code: zod.ZodIssueCode.custom,
+          message: 'Time is outside of restaurant operating hours',
+          path: ['time'],
+        });
+      }
+      
+      if (restaurant.blacklist.includes(data.phone)) {
+        context.addIssue({
+          code: zod.ZodIssueCode.custom,
+          message: 'Internal error, please try again later',
+          path: ['phone'],
+        });
+      }
     });
 };
