+++ frontend/serverless.yaml
@@ -46,55 +46,6 @@ functions:
       EMAIL_FROM: ${env:EMAIL_FROM}
       EMAIL_TO: ${env:EMAIL_TO}
 
-  createCheckoutSession:
-    handler: lambda/create-checkout-session.handler
-    memorySize: 1024
-    timeout: 28
-    role: ServerlessNextjsLambdaRole
-    events:
-      - http:
-          path: api/create-checkout-session
-          method: post
-          cors:
-            origin: ${env:ALLOW_ORIGIN} # Allow all domains (replace with specific domains in production)
-            headers:
-              - Content-Type
-              - Authorization
-              - X-Amz-Date
-              - X-Api-Key
-              - X-Amz-Security-Token
-              - X-Amz-User-Agent
-    environment:
-      STRIPE_SECRET_KEY: ${env:STRIPE_SECRET_KEY}
-      CLIENT_BASE_URL: ${env:CLIENT_BASE_URL}
-      ALLOW_ORIGIN: ${env:ALLOW_ORIGIN}
-
-  webhookHandler:
-    handler: lambda/webhook-handler.handler
-    memorySize: 1024
-    timeout: 28
-    role: ServerlessNextjsLambdaRole
-    events:
-      - http:
-          path: api/webhooks/stripe
-          method: post
-          cors:
-            origin: '*' 
-            headers:
-              - Content-Type
-              - Stripe-Signature
-              - Authorization
-              - X-Amz-Date
-              - X-Api-Key
-              - X-Amz-Security-Token
-              - X-Amz-User-Agent
-    environment:
-      STRIPE_SECRET_KEY: ${env:STRIPE_SECRET_KEY}
-      STRIPE_WEBHOOK_SECRET: ${env:STRIPE_WEBHOOK_SECRET}
-      SANITY_PROJECT_ID: ${env:SANITY_PROJECT_ID}
-      SANITY_DATASET: ${env:SANITY_DATASET}
-      SANITY_API_TOKEN: ${env:SANITY_API_TOKEN}
-
   createReservation:
     handler: lambda/create-reservation.handler
     memorySize: 1024
